requirements.txt
asyncua
RPi.GPIO